Refactore: move syncGroups to recentGroupsStorage.ts
